[refactor] スーパー、スーパーリファクタリング
