+++ app/components/StartScreen.tsx
@@ -1,4 +1,5 @@
-import { AddVideoButton } from "~/components/video-controller/AddVideoButton";
+import { JSX } from "react";
+import { AddVideoButton } from "~/components/controller/video-controller/AddVideoButton";
 
 import "~/styles/start-screen.css";
