+++ app/components/controller/PlayControl.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from "jotai";
-import { useEffect, useState } from "react";
+import { JSX, useEffect, useState } from "react";
 import { playerModelAtom, playStateAtom } from "~/atoms";
 import { PauseIcon, PlayIcon } from "~/components/common/icons";
 import { VideoDataModel } from "~/models/videoDataModel";
