+++ app/components/controller/VolumeControl.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from "jotai";
-import { useEffect, useState } from "react";
+import { JSX, useEffect, useState } from "react";
 import { muteStateAtom, playerModelAtom, volumeStateAtom } from "~/atoms";
 import { MuteIcon, VolumeLevel1Icon, VolumeLevel2Icon } from "~/components/common/icons";
 import { VideoDataModel } from "~/models/videoDataModel";
@@ -46,7 +46,7 @@ export function VolumeControl({ data }: { data: VideoDataModel }): JSX.Element {
 				type="button"
 				onClick={toggleMuted}>
 				{
-					isMuted ? <MuteIcon /> : (volume <= LOW_VOLUME_THRESHOLD ? <VolumeLevel1Icon /> : <VolumeLevel2Icon />)
+					isMuted ? <MuteIcon /> : (volume <= 0.5 ? <VolumeLevel1Icon /> : <VolumeLevel2Icon />)
 				}
 			</button>
