+++ app/components/controller/global-controller/GlobalMuteControl.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from "jotai";
-import { useEffect, useState } from "react";
+import { JSX, useEffect, useState } from "react";
 import { allMuteStateAtom, allPlayerModelsAtom } from "~/atoms";
 import { MuteIcon, VolumeLevel2Icon } from "~/components/common/icons";
