+++ app/components/controller/global-controller/GlobalPlayControl.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from "jotai";
-import { useEffect, useState } from "react";
+import { JSX, useEffect, useState } from "react";
 import { allPlayerModelsAtom, allPlayStateAtom } from "~/atoms";
 import { PauseIcon, PlayIcon } from "~/components/common/icons";
