--- app/components/video_controllers.css
@@ -1,182 +1,0 @@
-:root {
-	--background-color: #F5F5F5;
-	--background-gradient: linear-gradient(to bottom right var(--background-color), #F1F1F1);
-	--background-color-hover: #D6D6D6;
-	--background-color-active: #B7B7B7;
-	--widget-color-primary: #D9D9D9;
-	--icon-color-primary: #5A5A5A;
-}
-
-.video_controllers_container {
-	width: 100%;
-	height: fit-content;
-	display: flex;
-	flex-direction: column;
-	row-gap: 25px;
-	overflow-y: auto;
-	padding: 30px;
-
-	& .video_controller {
-		width: 100%;
-		height: fit-content;
-		padding: 17px 20px;
-		box-sizing: border-box;
-		border-radius: 10px;
-		display: flex;
-		flex-direction: column;
-		row-gap: 10px;
-		box-shadow: 3px 3px 4px #0000004D, -3px -3px 4px #FFFFFFB8;
-		transition: all 0.2s ease;
-		background: var(--background-gradient);
-
-		& .labels {
-			display: flex;
-			column-gap: 10px;
-
-			& .icon {
-				width: 50px;
-				height: 50px;
-				border-radius: 50%;
-				background-color: var(--widget-color-primary);
-			}
-
-			& .url {
-				flex-grow: 1;
-				height: 50px;
-				padding: 5px;
-				border-radius: 0 8px 8px 0;
-				background-color: var(--widget-color-primary);
-			}
-		}
-
-		& .controls {
-			display: flex;
-
-			& .control_item {
-				width: fit-content;
-				height: fit-content;
-				display: flex;
-				column-gap: 5px;
-				align-items: center;
-
-				& .control_button {
-					width: 50px;
-					height: 50px;
-					display: flex;
-					align-items: center;
-					justify-content: center;
-					border-radius: 50%;
-					background-size: 50%;
-					transition: all 0.2s ease;
-
-					&:hover {
-						background-color: var(--background-color-hover);
-					}
-
-					&:active {
-						background-color: var(--background-color-active);
-					}
-
-					&.play_button svg {
-						/* 微調整 */
-						transform: translateX(2px);
-					}
-				}
-
-				&.volume_control {
-					cursor: pointer;
-
-					& .volume_slider {
-						width: 0;
-						opacity: 0.0;
-						transition: all 0.1s ease;
-						appearance: none;
-						height: 6px;
-						border-radius: 99px;
-						background: linear-gradient(90deg, var(--icon-color-primary) calc(var(--ratio) * 100%), #dddddd calc(var(--ratio) * 100%));
-						box-shadow: inset 2px 2px 4px rgba(0, 0, 0, 0.25), inset -2px -2px 4px rgba(255, 255, 255, 0.15);
-						cursor: pointer;
-
-						&::-webkit-slider-thumb {
-							-webkit-appearance: none;
-							appearance: none;
-							width: 12px;
-							height: 12px;
-							border-radius: 50%;
-							background: var(--icon-color-primary);
-							box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.25), -2px -2px 4px rgba(255, 255, 255, 0.15);
-						}
-					}
-
-					&:hover .volume_slider {
-						width: 100px;
-						opacity: 1.0;
-					}
-				}
-
-				&.remove_control {
-					margin-left: auto;
-				}
-			}
-		}
-
-		&:hover {
-			scale: 1.025;
-			box-shadow: 1px 1px 8px #00000040, -1px -1px 8px #FFFFFFB8;
-		}
-	}
-
-	& .add_video_button_container {
-		width: 100%;
-		height: fit-content;
-		display: flex;
-		justify-content: center;
-
-		& .add_video_button {
-			width: 80px;
-			height: 50px;
-			border-radius: 99px;
-			background: var(--background-gradient);
-			box-shadow: 5px 5px 7px rgba(0, 0, 0, 0.25), -5px -5px 7px rgba(255, 255, 255, 0.6);
-			display: flex;
-			justify-content: center;
-			font-size: 50px;
-			display: flex;
-			align-items: center;
-			transition: all 0.3s ease;
-
-			&::before {
-				width: 30px;
-				height: 30px;
-				background: linear-gradient(to bottom right, #faf623, #f36bca);
-				background-size: 200% 200%;
-				clip-path: url(#plus_clip_icon);
-				content: "";
-				animation-name: gradation;
-				animation-duration: 7.0s;
-				animation-iteration-count: infinite;
-				animation-direction: alternate;
-				animation-timing-function: ease-in-out;
-			}
-
-			&:hover {
-				scale: 1.1;
-				box-shadow: 3px 3px 10px rgba(0, 0, 0, 0.25), -3px -3px 10px rgba(255, 255, 255, 0.6);
-
-				&::before {
-					animation-duration: 2.0s;
-				}
-			}
-		}
-	}
-}
-
-@keyframes gradation {
-	0% {
-		background-position: 0% 0%;
-	}
-
-	100% {
-		background-position: 100% 100%;
-	}
-}
