+++ app/styles/sidebar.css
